Move git stuff into its own namespace
